requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
pydantic==2.9.2
httpx==0.27.2
pytest==8.3.3
azure-identity==1.17.1